crud aula 33 - dialog de confirmação

+++ crud-angular-spring/crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 
 import { Course } from '../../model/course';
@@ -54,16 +55,24 @@ export class CoursesComponent {
   }
 
   onDelete(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh()
-        this.snackBar.open('Curso removido com sucesso.', 'X', { 
-          duration: 4000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'  
-        })
-      }, 
-      () => this.onError('Erro ao tentar remover curso.')
-    )
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh()
+            this.snackBar.open('Curso removido com sucesso.', 'X', { 
+              duration: 4000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'  
+            })
+          }, 
+          () => this.onError('Erro ao tentar remover curso.')
+        )
+      }
+    });
   }
 }
